pyproject.toml
[tool.poetry]
name = "redmineapi"
version = "0.1.0"
description = ""
authors = ["RomanSiu <siusiailoroman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
python-redmine = "^2.5.0"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
